+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The main reasoning behind this choice is the software is not built to be extende
 
 ## [Unreleased](https://github.com/miraheze/TSPortal/compare/v1...master)
 
+### Added
+
+- Add new filtering options for investigations and reports.
+
 ### Changed
 
 - Modified how buttons and the nav bar displays on smaller screens to make them more accessible.